Bugfixes and comments
Fixed a bug in tanks' OnLoad (the pole wasn't loaded correctly).
Fixes the selection of the leak resources in tanks (the last resource
could never be selected).
Fixes a bug during maintenance (the maintenance would fail if the perks
DID checkout, a not was missing).
Changed the perks node name to be DANGIT_PERKS in all occasions (to
match the node returned by the generator).
Renamed InspectionName to ScreenName.
Updated .version file